pyproject.toml
[project]
name = "mining-industries-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-auth>=2.39.0",
    "gspread>=6.2.0",
    "jupyterlab>=4.4.1",
    "pandas>=2.2.3",
    "peewee>=3.17.9",
    "tabulate>=0.9.0",
]
